backend/requirements.txt
fastapi
uvicorn
playwright
python-docx
pandas
openpyxl
python-multipart
supabase
python-dotenv
